+++ filament/backend/src/metal/MetalDriver.mm
@@ -62,29 +62,33 @@ MetalDriver::MetalDriver(backend::MetalPlatform* platform) noexcept
     mContext->device = mPlatform.createDevice();
     assert_invariant(mContext->device);
 
+    initializeSupportedGpuFamilies(mContext);
+
+    utils::slog.d << "Supported GPU families: " << utils::io::endl;
+    if (mContext->highestSupportedGpuFamily.common > 0) {
+        utils::slog.d << "  MTLGPUFamilyCommon" << (int) mContext->highestSupportedGpuFamily.common << utils::io::endl;
+    }
+    if (mContext->highestSupportedGpuFamily.apple > 0) {
+        utils::slog.d <<   "MTLGPUFamilyApple" << (int) mContext->highestSupportedGpuFamily.apple << utils::io::endl;
+    }
+    if (mContext->highestSupportedGpuFamily.mac > 0) {
+        utils::slog.d << "  MTLGPUFamilyMac" << (int) mContext->highestSupportedGpuFamily.mac << utils::io::endl;
+    }
+
     // In order to support texture swizzling, the GPU needs to support it and the system be running
     // macOS 10.15+ / iOS 13+.
     mContext->supportsTextureSwizzling = false;
     if (@available(macOS 10.15, iOS 13, *)) {
-#if defined(IOS)
         mContext->supportsTextureSwizzling =
-                [mContext->device supportsFeatureSet:MTLFeatureSet_iOS_GPUFamily1_v1];
-#else
-        mContext->supportsTextureSwizzling =
-                [mContext->device supportsFeatureSet:MTLFeatureSet_macOS_GPUFamily2_v1];
-#endif
+            mContext->highestSupportedGpuFamily.apple >= 1 ||   // all Apple GPUs
+            mContext->highestSupportedGpuFamily.mac   >= 2;     // newer macOS GPUs
     }
 
     mContext->maxColorRenderTargets = 4;
-#if defined(IOS)
-    if ([mContext->device supportsFeatureSet:MTLFeatureSet_iOS_GPUFamily2_v1]) {
+    if (mContext->highestSupportedGpuFamily.apple >= 2 ||
+        mContext->highestSupportedGpuFamily.mac >= 1) {
         mContext->maxColorRenderTargets = 8;
     }
-#else
-    if ([mContext->device supportsFeatureSet:MTLFeatureSet_macOS_GPUFamily1_v1]) {
-        mContext->maxColorRenderTargets = 8;
-    }
-#endif
 
     // Round requested sample counts down to the nearest device-supported sample count.
     auto& sc = mContext->sampleCountLookup;
@@ -254,7 +258,7 @@ void MetalDriver::importTextureR(Handle<HwTexture> th, intptr_t i,
     ASSERT_PRECONDITION(metalTexture.mipmapLevelCount == levels,
             "Imported id<MTLTexture> levels (%d) != Filament texture levels (%d)",
             metalTexture.mipmapLevelCount, levels);
-    MTLPixelFormat filamentMetalFormat = getMetalFormat(format);
+    MTLPixelFormat filamentMetalFormat = getMetalFormat(mContext, format);
     ASSERT_PRECONDITION(metalTexture.pixelFormat == filamentMetalFormat,
             "Imported id<MTLTexture> format (%d) != Filament texture format (%d)",
             metalTexture.pixelFormat, filamentMetalFormat);
@@ -290,7 +294,7 @@ void MetalDriver::createRenderTargetR(Handle<HwRenderTarget> rth,
     auto& sc = mContext->sampleCountLookup;
     samples = sc[std::min(MAX_SAMPLE_COUNT, samples)];
 
-    MetalRenderTarget::Attachment colorAttachments[MRT::MAX_SUPPORTED_RENDER_TARGET_COUNT] = {{ nil }};
+    MetalRenderTarget::Attachment colorAttachments[MRT::MAX_SUPPORTED_RENDER_TARGET_COUNT] = {{}};
     for (size_t i = 0; i < MRT::MAX_SUPPORTED_RENDER_TARGET_COUNT; i++) {
         const auto& buffer = color[i];
         if (!buffer.handle) {
@@ -303,20 +307,16 @@ void MetalDriver::createRenderTargetR(Handle<HwRenderTarget> rth,
         ASSERT_PRECONDITION(colorTexture->texture,
                 "Color texture passed to render target has no texture allocation");
         colorTexture->updateLodRange(buffer.level);
-        colorAttachments[i].texture = colorTexture->texture;
-        colorAttachments[i].level = color[i].level;
-        colorAttachments[i].layer = color[i].layer;
+        colorAttachments[i] = { colorTexture, color[i].level, color[i].layer };
     }
 
-    MetalRenderTarget::Attachment depthAttachment = { nil };
+    MetalRenderTarget::Attachment depthAttachment = {};
     if (depth.handle) {
         auto depthTexture = handle_cast<MetalTexture>(depth.handle);
         ASSERT_PRECONDITION(depthTexture->texture,
                 "Depth texture passed to render target has no texture allocation.");
         depthTexture->updateLodRange(depth.level);
-        depthAttachment.texture = depthTexture->texture;
-        depthAttachment.level = depth.level;
-        depthAttachment.layer = depth.layer;
+        depthAttachment = { depthTexture, depth.level, depth.layer };
     }
     ASSERT_POSTCONDITION(!depth.handle || any(targetBufferFlags & TargetBufferFlags::DEPTH),
             "The DEPTH flag was specified, but no depth texture provided.");
@@ -598,7 +598,7 @@ FenceStatus MetalDriver::wait(Handle<HwFence> fh, uint64_t timeout) {
 }
 
 bool MetalDriver::isTextureFormatSupported(TextureFormat format) {
-    return MetalTexture::decidePixelFormat(mContext->device, format) != MTLPixelFormatInvalid;
+    return MetalTexture::decidePixelFormat(mContext, format) != MTLPixelFormatInvalid;
 }
 
 bool MetalDriver::isTextureSwizzleSupported() {
@@ -608,7 +608,7 @@ bool MetalDriver::isTextureSwizzleSupported() {
 bool MetalDriver::isTextureFormatMipmappable(TextureFormat format) {
     // Derived from the Metal 3.0 Feature Set Tables.
     // In order for a format to be mipmappable, it must be color-renderable and filterable.
-    MTLPixelFormat metalFormat = MetalTexture::decidePixelFormat(mContext->device, format);
+    MTLPixelFormat metalFormat = MetalTexture::decidePixelFormat(mContext, format);
     switch (metalFormat) {
         // Mipmappable across all devices:
         case MTLPixelFormatR8Unorm:
@@ -645,7 +645,7 @@ bool MetalDriver::isTextureFormatMipmappable(TextureFormat format) {
 }
 
 bool MetalDriver::isRenderTargetFormatSupported(TextureFormat format) {
-    MTLPixelFormat mtlFormat = getMetalFormat(format);
+    MTLPixelFormat mtlFormat = getMetalFormat(mContext, format);
     // RGB9E5 isn't supported on Mac as a color render target.
     return mtlFormat != MTLPixelFormatInvalid && mtlFormat != MTLPixelFormatRGB9E5Float;
 }
@@ -948,7 +948,7 @@ void MetalDriver::readPixels(Handle<HwRenderTarget> src, uint32_t x, uint32_t y,
     auto srcTarget = handle_cast<MetalRenderTarget>(src);
     // We always readPixels from the COLOR0 attachment.
     MetalRenderTarget::Attachment color = srcTarget->getDrawColorAttachment(0);
-    id<MTLTexture> srcTexture = color.texture;
+    id<MTLTexture> srcTexture = color.getTexture();
     size_t miplevel = color.level;
 
     auto chooseMetalPixelFormat = [] (PixelDataFormat format, PixelDataType type) {
@@ -1078,12 +1078,12 @@ void MetalDriver::blit(TargetBufferFlags buffers,
         MetalRenderTarget::Attachment dstColorAttachment = dstTarget->getDrawColorAttachment(0);
 
         if (srcColorAttachment && dstColorAttachment) {
-            ASSERT_PRECONDITION(isBlitableTextureType(srcColorAttachment.texture.textureType) &&
-                                isBlitableTextureType(dstColorAttachment.texture.textureType),
+            ASSERT_PRECONDITION(isBlitableTextureType(srcColorAttachment.getTexture().textureType) &&
+                                isBlitableTextureType(dstColorAttachment.getTexture().textureType),
                                "Metal does not support blitting to/from non-2D textures.");
 
-            args.source.color = srcColorAttachment.texture;
-            args.destination.color = dstColorAttachment.texture;
+            args.source.color = srcColorAttachment.getTexture();
+            args.destination.color = dstColorAttachment.getTexture();
             args.source.level = srcColorAttachment.level;
             args.destination.level = dstColorAttachment.level;
         }
@@ -1094,12 +1094,12 @@ void MetalDriver::blit(TargetBufferFlags buffers,
         MetalRenderTarget::Attachment dstDepthAttachment = dstTarget->getDepthAttachment();
 
         if (srcDepthAttachment && dstDepthAttachment) {
-            ASSERT_PRECONDITION(isBlitableTextureType(srcDepthAttachment.texture.textureType) &&
-                                isBlitableTextureType(dstDepthAttachment.texture.textureType),
+            ASSERT_PRECONDITION(isBlitableTextureType(srcDepthAttachment.getTexture().textureType) &&
+                                isBlitableTextureType(dstDepthAttachment.getTexture().textureType),
                                "Metal does not support blitting to/from non-2D textures.");
 
-            args.source.depth = srcDepthAttachment.texture;
-            args.destination.depth = dstDepthAttachment.texture;
+            args.source.depth = srcDepthAttachment.getTexture();
+            args.destination.depth = dstDepthAttachment.getTexture();
 
             if (args.blitColor()) {
                 // If blitting color, we've already set the source and destination levels.
@@ -1143,12 +1143,12 @@ void MetalDriver::draw(backend::PipelineState ps, Handle<HwRenderPrimitive> rph)
         if (!attachment) {
             continue;
         }
-        colorPixelFormat[i] = attachment.texture.pixelFormat;
+        colorPixelFormat[i] = attachment.getPixelFormat();
     }
     MTLPixelFormat depthPixelFormat = MTLPixelFormatInvalid;
     const auto& depthAttachment = mContext->currentRenderTarget->getDepthAttachment();
     if (depthAttachment) {
-        depthPixelFormat = depthAttachment.texture.pixelFormat;
+        depthPixelFormat = depthAttachment.getPixelFormat();
     }
     metal::PipelineState pipelineState {
         .vertexFunction = program->vertexFunction,