+++ filament/backend/src/metal/MetalHandles.h
@@ -29,13 +29,13 @@
 #include "MetalExternalImage.h"
 #include "MetalState.h" // for MetalState::VertexDescription
 
+#include <utils/FixedCapacityVector.h>
 #include <utils/Panic.h>
 
 #include <atomic>
 #include <condition_variable>
 #include <memory>
 #include <type_traits>
-#include <vector>
 
 namespace filament {
 namespace backend {
@@ -127,7 +127,7 @@ struct MetalVertexBuffer : public HwVertexBuffer {
     MetalVertexBuffer(MetalContext& context, uint8_t bufferCount, uint8_t attributeCount,
             uint32_t vertexCount, AttributeArray const& attributes);
 
-    std::vector<MetalBuffer*> buffers;
+    utils::FixedCapacityVector<MetalBuffer*> buffers;
 };
 
 struct MetalIndexBuffer : public HwIndexBuffer {
@@ -194,11 +194,16 @@ struct MetalTexture : public HwTexture {
             const PixelBufferShape& shape);
     void updateLodRange(uint32_t level);
 
-    static MTLPixelFormat decidePixelFormat(id<MTLDevice> device, TextureFormat format);
+    static MTLPixelFormat decidePixelFormat(MetalContext* context, TextureFormat format);
 
     MetalContext& context;
     MetalExternalImage externalImage;
     id<MTLTexture> texture = nil;
+
+    // A "sidecar" texture used to implement automatic MSAA resolve.
+    // This is created by MetalRenderTarget and stored here so it can be used with multiple
+    // render targets.
+    id<MTLTexture> msaaSidecar = nil;
 
     // If non-nil, a swizzled texture view to use instead of "texture".
     // Filament swizzling only affects texture reads, so this should not be used when the texture is
@@ -217,14 +222,46 @@ struct MetalSamplerGroup : public HwSamplerGroup {
 class MetalRenderTarget : public HwRenderTarget {
 public:
 
-    struct Attachment {
-        id<MTLTexture> texture = nil;
-        uint8_t level = 0;
-        uint16_t layer = 0;
+    class Attachment {
+    public:
+
+        friend class MetalRenderTarget;
+
+        Attachment() = default;
+        Attachment(MetalTexture* metalTexture, uint8_t level = 0, uint16_t layer = 0) :
+                level(level), layer(layer),
+                texture(metalTexture->texture),
+                metalTexture(metalTexture) { }
+
+        id<MTLTexture> getTexture() const {
+            return texture;
+        }
+
+        NSUInteger getSampleCount() const {
+            return texture ? texture.sampleCount : 0u;
+        }
+
+        MTLPixelFormat getPixelFormat() const {
+            return texture ? texture.pixelFormat : MTLPixelFormatInvalid;
+        }
 
         explicit operator bool() const {
             return texture != nil;
         }
+
+        uint8_t level = 0;
+        uint16_t layer = 0;
+
+    private:
+
+        id<MTLTexture> getMSAASidecarTexture() const {
+            // This should only be called from render targets associated with a MetalTexture.
+            assert_invariant(metalTexture);
+            return metalTexture->msaaSidecar;
+        }
+
+        id<MTLTexture> texture = nil;
+        MetalTexture* metalTexture = nullptr;
     };
 
     MetalRenderTarget(MetalContext* context, uint32_t width, uint32_t height, uint8_t samples,
@@ -254,10 +291,6 @@ private:
 
     Attachment color[MRT::MAX_SUPPORTED_RENDER_TARGET_COUNT] = {};
     Attachment depth = {};
-
-    // "Sidecar" textures used to implement automatic MSAA resolve.
-    id<MTLTexture> multisampledColor[MRT::MAX_SUPPORTED_RENDER_TARGET_COUNT] = { 0 };
-    id<MTLTexture> multisampledDepth = nil;
 };
 
 // MetalFence is used to implement both Fences and Syncs.