+++ filament/include/filament/Camera.h
@@ -207,6 +207,9 @@ public:
      *       0 0 tz c              0 0 c tz
      *       0 0 -1 0              0 0 0 1
      *
+     * The projection matrix must define an NDC system that must match the OpenGL convention,
+     * that is all 3 axis are mapped to [-1, 1].
+     *
      * @param projection  custom projection matrix used for rendering and culling
      * @param near        distance in world units from the camera to the near plane. \p near > 0.
      * @param far         distance in world units from the camera to the far plane. \p far > \p near.
@@ -220,6 +223,9 @@ public:
      *       0 b ty 0              0 b 0 ty
      *       0 0 tz c              0 0 c tz
      *       0 0 -1 0              0 0 0 1
+     *
+     * The projection matrices must define an NDC system that must match the OpenGL convention,
+     * that is all 3 axis are mapped to [-1, 1].
      *
      * @param projection  custom projection matrix used for rendering
      * @param projectionForCulling  custom projection matrix used for culling