+++ filament/src/Scene.cpp
@@ -140,6 +140,7 @@ void FScene::prepare(const mat4f& worldOriginTransform, bool shadowReceiversAreC
                     worldAABB.center,               // WORLD_AABB_CENTER
                     0,                              // VISIBLE_MASK
                     rcm.getMorphWeights(ri),        // MORPH_WEIGHTS
+                    rcm.getChannels(ri),            // CHANNELS
                     rcm.getLayerMask(ri),           // LAYERS
                     worldAABB.halfExtent,           // WORLD_AABB_EXTENT
                     {},                             // PRIMITIVES
@@ -239,25 +240,26 @@ void FScene::updateUBOs(utils::Range<uint32_t> visibleRenderables, backend::Hand
 
         FRenderableManager::Visibility visibility = sceneData.elementAt<VISIBILITY_STATE>(i);
         hasContactShadows = hasContactShadows || visibility.screenSpaceContactShadows;
-        UniformBuffer::setUniform(buffer,
-                offset + offsetof(PerRenderableUib, skinningEnabled),
-                uint32_t(visibility.skinning));
 
         UniformBuffer::setUniform(buffer,
-                offset + offsetof(PerRenderableUib, morphingEnabled),
-                uint32_t(visibility.morphing));
-
-        UniformBuffer::setUniform(buffer,
-                offset + offsetof(PerRenderableUib, screenSpaceContactShadows),
-                uint32_t(visibility.screenSpaceContactShadows));
+                offset + offsetof(PerRenderableUib, flags),
+                PerRenderableUib::packFlags(
+                        visibility.skinning,
+                        visibility.morphing,
+                        visibility.screenSpaceContactShadows));
 
         UniformBuffer::setUniform(buffer,
                 offset + offsetof(PerRenderableUib, morphWeights),
                 sceneData.elementAt<MORPH_WEIGHTS>(i));
 
+        UniformBuffer::setUniform(buffer,
+                offset + offsetof(PerRenderableUib, channels),
+                (uint32_t)sceneData.elementAt<CHANNELS>(i));
+
         // TODO: We need to find a better way to provide the scale information per object
         UniformBuffer::setUniform(buffer,
-                offset + offsetof(PerRenderableUib, userData), sceneData.elementAt<USER_DATA>(i));
+                offset + offsetof(PerRenderableUib, userData),
+                sceneData.elementAt<USER_DATA>(i));
     }
 
     // TODO: handle static objects separately
@@ -305,11 +307,17 @@ void FScene::prepareDynamicLights(const CameraInfo& camera, ArenaScope& rootAren
         const size_t gpuIndex = i - DIRECTIONAL_LIGHTS_COUNT;
         auto li = instances[i];
         lp[gpuIndex].positionFalloff      = { spheres[i].xyz, lcm.getSquaredFalloffInv(li) };
-        lp[gpuIndex].colorIntensity       = { lcm.getColor(li), lcm.getIntensity(li) };
+        lp[gpuIndex].color                = { lcm.getColor(li), 0.0f };
         lp[gpuIndex].directionIES         = { directions[i], 0.0f };
         lp[gpuIndex].spotScaleOffset      = lcm.getSpotParams(li).scaleOffset;
-        lp[gpuIndex].shadow               = { shadowInfo[i].pack() };
-        lp[gpuIndex].type                 = lcm.isPointLight(li) ? 0u : 1u;
+        lp[gpuIndex].intensity            = lcm.getIntensity(li);
+        lp[gpuIndex].typeShadow           = LightsUib::packTypeShadow(
+                lcm.isPointLight(li) ? 0u : 1u,
+                shadowInfo[i].contactShadows,
+                shadowInfo[i].index,
+                shadowInfo[i].layer);
+        lp[gpuIndex].channels             = LightsUib::packChannels(lcm.getLightChannels(li), shadowInfo[i].castsShadows);
+        lp[gpuIndex].reserved             = {};
     }
 
     driver.updateBufferObject(lightUbh, { lp, positionalLightCount * sizeof(LightsUib) }, 0);