+++ filament/src/View.cpp
@@ -355,6 +355,7 @@ void FView::prepareLighting(FEngine& engine, FEngine::DriverApi& driver, ArenaSc
 
         s.lightDirection = l;
         s.lightColorIntensity = colorIntensity;
+        s.lightChannels = lcm.getLightChannels(directionalLight);
 
         const bool isSun = lcm.isSunLight(directionalLight);
         // The last parameter must be < 0.0f for regular directional lights
@@ -682,10 +683,12 @@ void FView::prepareSSAO(Handle<HwTexture> ssao) const noexcept {
                          SamplerMagFilter::LINEAR : SamplerMagFilter::NEAREST
     });
 
-    const float edgeDistance = 1.0 / 0.0625;// TODO: don't hardcode this
+    const float edgeDistance = 1.0f / mAmbientOcclusionOptions.bilateralThreshold;
     auto& s = mPerViewUb.edit();
     s.aoSamplingQualityAndEdgeDistance =
             mAmbientOcclusionOptions.enabled && highQualitySampling ? edgeDistance : 0.0f;
+    s.aoBentNormals =
+            mAmbientOcclusionOptions.enabled && mAmbientOcclusionOptions.bentNormals ? 1.0f : 0.0f;
 }
 
 void FView::prepareSSR(backend::Handle<backend::HwTexture> ssr, float refractionLodOffset) const noexcept {