+++ shaders/src/light_punctual.fs
@@ -30,9 +30,19 @@ uvec3 getFroxelCoords(const highp vec3 fragCoords) {
     froxelCoord.xy = uvec2(fragCoords.xy * frameUniforms.resolution.xy *
             vec2(frameUniforms.oneOverFroxelDimension, frameUniforms.oneOverFroxelDimensionY));
 
-    froxelCoord.z = uint(max(0.0,
-            log2(frameUniforms.zParams.x * fragCoords.z + frameUniforms.zParams.y) *
-                    frameUniforms.zParams.z + frameUniforms.zParams.w));
+    // go from screen-space (non-inverted) Z to normalized view-space Z (i.e. scaled by 1/zLightFar)
+    // (see Froxelizer.cpp)
+    highp float viewSpaceNormalizedZ = frameUniforms.zParams.x * fragCoords.z + frameUniforms.zParams.y;
+
+    // frameUniforms.zParams.w is actually the number of z-slices, make sure it's mediump
+    float zSliceCount = frameUniforms.zParams.w;
+
+    // compute the sliceZ mapping in highp, store in mediump
+    float sliceZWithoutOffset = log2(viewSpaceNormalizedZ) * frameUniforms.zParams.z;
+
+    // finally discretize the mapping into slices
+    // We need to clamp because the far plane (z=1) is out of bounds, any smaller z is not.
+    froxelCoord.z = uint(clamp(sliceZWithoutOffset + zSliceCount, 0.0, zSliceCount - 1.0));
 
     return froxelCoord;
 }
@@ -113,13 +123,26 @@ float getAngleAttenuation(const vec3 lightDir, const vec3 l, const vec2 scaleOff
  * The light parameters used to compute the Light structure are fetched from the
  * lightsUniforms uniform buffer.
  */
+
 Light getLight(const uint index) {
     // retrieve the light data from the UBO
     uint lightIndex = getLightIndex(index);
-    highp vec4 positionFalloff       = lightsUniforms.lights[lightIndex][0];
-    highp vec4 colorIntensity        = lightsUniforms.lights[lightIndex][1];
-          vec4 directionIES          = lightsUniforms.lights[lightIndex][2];
-    highp vec4 scaleOffsetShadowType = lightsUniforms.lights[lightIndex][3];
+
+    highp mat4 data = lightsUniforms.lights[lightIndex];
+
+    highp vec4 positionFalloff = data[0];
+    vec4 color = vec4(
+        unpackHalf2x16(floatBitsToUint(data[1][0])),
+        unpackHalf2x16(floatBitsToUint(data[1][1]))
+    );
+    vec4 directionIES = vec4(
+        unpackHalf2x16(floatBitsToUint(data[1][2])),
+        unpackHalf2x16(floatBitsToUint(data[1][3]))
+    );
+    vec2 scaleOffset = unpackHalf2x16(floatBitsToUint(data[2][0]));
+    highp float intensity = data[2][1];
+    highp uint typeShadow = floatBitsToUint(data[2][2]);
+    highp uint channels = floatBitsToUint(data[2][3]);
 
     // poition-to-light vector
     highp vec3 worldPosition = vertex_worldPosition;
@@ -127,8 +150,8 @@ Light getLight(const uint index) {
 
     // and populate the Light structure
     Light light;
-    light.colorIntensity.rgb = colorIntensity.rgb;
-    light.colorIntensity.w = computePreExposedIntensity(colorIntensity.w, frameUniforms.exposure);
+    light.colorIntensity.rgb = color.rgb;
+    light.colorIntensity.w = computePreExposedIntensity(intensity, frameUniforms.exposure);
     light.l = normalize(posToLight);
     light.attenuation = getDistanceAttenuation(posToLight, positionFalloff.w);
     light.NoL = saturate(dot(shading_normal, light.l));
@@ -137,15 +160,15 @@ Light getLight(const uint index) {
     light.contactShadows = false;
     light.shadowIndex = 0u;
     light.shadowLayer = 0u;
+    light.channels = channels;
 
-    uint type = floatBitsToUint(scaleOffsetShadowType.w);
+    uint type = typeShadow & 0x1u;
     if (type == LIGHT_TYPE_SPOT) {
-        light.attenuation *= getAngleAttenuation(-directionIES.xyz, light.l, scaleOffsetShadowType.xy);
-        uint shadowBits = floatBitsToUint(scaleOffsetShadowType.z);
-        light.castsShadows = bool(shadowBits & 0x1u);
-        light.contactShadows = bool((shadowBits >> 1u) & 0x1u);
-        light.shadowIndex = (shadowBits >> 2u) & 0xFu;
-        light.shadowLayer = (shadowBits >> 6u) & 0xFu;
+        light.attenuation *= getAngleAttenuation(-directionIES.xyz, light.l, scaleOffset);
+        light.contactShadows = bool(typeShadow & 0x10u);
+        light.shadowIndex = (typeShadow >>  8u) & 0xFFu;
+        light.shadowLayer = (typeShadow >> 16u) & 0xFFu;
+        light.castsShadows   = bool(channels & 0x10000u);
     }
 
     return light;
@@ -171,10 +194,15 @@ void evaluatePunctualLights(const MaterialInputs material,
 
     uint index = froxel.recordOffset;
     uint end = index + froxel.count;
+    uint channels = objectUniforms.channels & 0xFFu;
 
     // Iterate point lights
     for ( ; index < end; index++) {
         Light light = getLight(index);
+        if ((light.channels & channels) == 0u) {
+            continue;
+        }
+
 #if defined(MATERIAL_CAN_SKIP_LIGHTING)
         if (light.NoL <= 0.0 || light.attenuation <= 0.0) {
             continue;
@@ -189,7 +217,7 @@ void evaluatePunctualLights(const MaterialInputs material,
                     getSpotLightSpacePosition(light.shadowIndex));
             }
             if (light.contactShadows && visibility > 0.0) {
-                if (objectUniforms.screenSpaceContactShadows != 0u) {
+                if ((objectUniforms.flags & FILAMENT_OBJECT_CONTACT_SHADOWS_BIT) != 0u) {
                     visibility *= 1.0 - screenSpaceContactShadow(light.l);
                 }
             }