+++ web/filament-js/jsbindings.cpp
@@ -365,7 +365,12 @@ value_object<filament::View::AmbientOcclusionOptions>("View$AmbientOcclusionOpti
     .field("resolution", &filament::View::AmbientOcclusionOptions::resolution)
     .field("intensity", &filament::View::AmbientOcclusionOptions::intensity)
     .field("bilateralThreshold", &filament::View::AmbientOcclusionOptions::bilateralThreshold)
-    .field("quality", &filament::View::AmbientOcclusionOptions::quality);
+    .field("quality", &filament::View::AmbientOcclusionOptions::quality)
+    .field("lowPassFilter", &filament::View::AmbientOcclusionOptions::lowPassFilter)
+    .field("upsampling", &filament::View::AmbientOcclusionOptions::upsampling)
+    .field("enabled", &filament::View::AmbientOcclusionOptions::enabled)
+    .field("bentNormals", &filament::View::AmbientOcclusionOptions::bentNormals)
+    .field("minHorizonAngleRad", &filament::View::AmbientOcclusionOptions::minHorizonAngleRad);
     // TODO: ssct options
 
 value_object<filament::View::DepthOfFieldOptions>("View$DepthOfFieldOptions")
@@ -975,6 +980,10 @@ class_<RenderableBuilder>("RenderableManager$Builder")
             (RenderableBuilder* builder, size_t index, uint16_t order), {
         return &builder->blendOrder(index, order); })
 
+    .BUILDER_FUNCTION("lightChannel", RenderableBuilder,
+            (RenderableBuilder* builder, unsigned int channel, bool enable), {
+        return &builder->lightChannel(channel, enable); })
+
     .function("_build", EMBIND_LAMBDA(int, (RenderableBuilder* builder,
             Engine* engine, utils::Entity entity), {
         return (int) builder->build(*engine, entity);
@@ -1001,6 +1010,8 @@ class_<RenderableManager>("RenderableManager")
     .function("setReceiveShadows", &RenderableManager::setReceiveShadows)
     .function("isShadowCaster", &RenderableManager::isShadowCaster)
     .function("isShadowReceiver", &RenderableManager::isShadowReceiver)
+    .function("setLightChannel", &RenderableManager::setLightChannel)
+    .function("getLightChannel", &RenderableManager::getLightChannel)
 
     .function("setBones", EMBIND_LAMBDA(void, (RenderableManager* self,
             RenderableManager::Instance instance, emscripten::val transforms, size_t offset), {
@@ -1139,7 +1150,10 @@ class_<LightBuilder>("LightManager$Builder")
     .BUILDER_FUNCTION("sunHaloSize", LightBuilder,
             (LightBuilder* builder, float value), { return &builder->sunHaloSize(value); })
     .BUILDER_FUNCTION("sunHaloFalloff", LightBuilder,
-            (LightBuilder* builder, float value), { return &builder->sunHaloFalloff(value); });
+            (LightBuilder* builder, float value), { return &builder->sunHaloFalloff(value); })
+    .BUILDER_FUNCTION("lightChannel", LightBuilder,
+            (LightBuilder* builder, unsigned int channel, bool enable), {
+        return &builder->lightChannel(channel, enable); });
 
 class_<LightManager>("LightManager")
     .function("hasComponent", &LightManager::hasComponent)
@@ -1186,6 +1200,8 @@ class_<LightManager>("LightManager")
     .function("getSunHaloFalloff", &LightManager::getSunHaloFalloff)
     .function("setShadowCaster", &LightManager::setShadowCaster)
     .function("isShadowCaster", &LightManager::isShadowCaster)
+    .function("setLightChannel", &LightManager::setLightChannel)
+    .function("getLightChannel", &LightManager::getLightChannel)
     ;
 
 /// LightManager$Instance ::class:: Component instance returned by [LightManager]